// frontend/src/entities/structure/modules/blocks/types/actions.ts
export enum ActionTypes {
    ADD_BLOCK = "ADD_BLOCK",
    ADD_BLOCK_BODY_TO_BLOCK = "ADD_BLOCK_BODY_TO_BLOCK",
    ADD_BLOCK_TITLE_TO_BLOCK = "ADD_BLOCK_TITLE_TO_BLOCK",
    REMOVE_BLOCK_BODY_FROM_BLOCK = "REMOVE_BLOCK_BODY_FROM_BLOCK",
    REMOVE_BLOCK_TITLE_FROM_BLOCK = "REMOVE_BLOCK_TITLE_FROM_BLOCK",
    ADD_BLOCK_HEADER_TO_BLOCK = "ADD_BLOCK_HEADER_TO_BLOCK",
    REMOVE_BLOCK_HEADER_FROM_BLOCK = "REMOVE_BLOCK_HEADER_FROM_BLOCK",
    ADD_BLOCK_HEADER_SECOND_ROW_TO_BLOCK = "ADD_BLOCK_HEADER_SECOND_ROW_TO_BLOCK",
    REMOVE_BLOCK_HEADER_SECOND_ROW_FROM_BLOCK = "REMOVE_BLOCK_HEADER_SECOND_ROW_FROM_BLOCK",
    CLEAR = "CLEAR",
}